# api-docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: CLM-EXT-Learning-objects API
  version: 1.0.0
  description: API endpoints the clm-ext-learning_objects module offers
servers:
- url: "{scheme}://{hostname}:{port}{path}"
  description: The production API server
  variables:
    hostname:
      default: localhost
    port:
      default: '5002'
    path:
      default: ''
    scheme:
      default: http
security:
- bearerAuth: []
paths:
...
  "/learningObjects/mgmt":
    get:
      tags:
      - mgmt-los
      summary: 'Retrieves all learning objects [Minimum Role: "Learner"]'
      description: Retrieve all learning objects the user has access to through the
        group enrollments.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      responses:
        '200':
          description: Successfully created learning object
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/LearningObject"
        '400':
          description: Bad request - Invalid input or validation error
        '401':
          description: Unauthorized
    post:
      tags:
      - mgmt-los
      summary: 'Create a Learning object [Minimum Role: "Instructor"]'
      description: Create a Learning object
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LearningObject"
      responses:
        '200':
          description: Successfully created learning object
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LearningObject"
        '400':
          description: Bad request - Invalid input or validation error
        '401':
          description: Unauthorized
  "/learningObjects/mgmt/{id}":
    put:
      tags:
      - mgmt-los
      summary: 'Update an existing Learning Object [Minimum Role: "Instructor"]'
      description: Modify an existing Learning object's attributes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LearningObject"
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      - in: path
        name: id
        description: The id of the learning object
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully updated learning object
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LearningObject"
        '400':
          description: Bad request - Invalid input or validation error
        '401':
          description: Unauthorized
    patch:
      tags:
      - mgmt-los
      summary: 'Update an existing Learning Object [Minimum Role: "Instructor"]'
      description: Modify an existing Learning object's attribute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LearningObject"
      parameters:
      - in: path
        name: id
        description: The id of the learning object
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully partially updated learning object
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LearningObject"
        '400':
          description: Bad request - Invalid input or validation error
        '401':
          description: Unauthorized
    delete:
      tags:
      - mgmt-los
      summary: 'Deletes an existing Learning Object [Minimum Role: "Instructor"]'
      description: Delete an existing learning object
      parameters:
      - in: path
        name: id
        description: The id of the learning object
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Successfully deleted learning object
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  "/learningObjects/mgmt/relations":
    get:
      tags:
      - mgmt-los
      summary: 'Get the relations from the learning objects the user has access to
        [Minimum Role : ''Learner'']'
      description: Retrieve relations between various learning objects that the authenticated
        user has permissions to view.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      responses:
        '200':
          description: A list of relations between learning objects.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/relation"
  "/learningObjects/mgmt/{id}/learningObjects/{childLoId}":
    post:
      tags:
      - mgmt-los
      summary: 'Add learning object to learning object [Minimum Role: ''Instructor'']'
      description: Associates a specific child learning object with another parent
        learning object, given their respective identifiers.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: childLoId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully associated the child learning object with the
            parent.
  "/learningObjects/mgmt/{id}/tools/{toolId}":
    post:
      tags:
      - mgmt-los
      summary: 'Add tool to learning object [Minimum Role: ''Instructor'']'
      description: Link a specific tool to a learning object by their respective identifiers.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: toolId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully added the tool to the learning object.
  "/learningObjects/mgmt/{id}/relations/{relationId}":
    put:
      tags:
      - mgmt-los
      summary: 'Change the order of a learning object within a learning object, or
        change the associated tool to a specific learning object [Minimum Role: ''Instructor'']'
      description: Modify the order of a specific learning object inside another or
        alter the tool linked to a given learning object.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: relationId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully updated the learning object's order or associated
            tool.
...
components:
  schemas:
    relation:
      type: object
      properties:
        fromType:
          type: string
          description: The type of the node
          default: fromTypeNode
        toType:
          type: string
          description: The type of the target node
          default: toTypeNode
        fromId:
          type: string
          description: The id of the node
          default: fromNodeId
        toId:
          type: string
          description: The id of the target node
          default: toNodeId
        order:
          type: number
          description: The order of the relation. Used for example ordering the enrollments
            of a group/user
          default: 0
    LearningObject:
      type: object
      required:
      - displayName
      properties:
        _id:
          type: string
          description: The id of the learning object
        displayName:
          type: string
          description: Display name for the learning object
          minLength: 5
        iconUrl:
          type: string
          description: Icon URL for the learning object
...
  parameters:
    accessToken:
      name: x-access-token
      in: header
      description: The access token
      required: true
      example: exampleAccessToken
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    refreshAuth:
      type: apiKey
      in: header
      name: x-refresh-token
...